requirements.txt
Django==4.2.7
django-ckeditor==6.7.0
django-filter==23.4
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.1.0
requests==2.31.0
setuptools==69.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
